requirements.txt
ollama
llama-cpp-python
groq
faiss-cpu
sentence-transformers
pandas
pydantic
python-dotenv
streamlit
plotly